Improve the interface for pull request summaries
This uses a Summary class with an `as_markdown` method to supply the
review for a given pull request.
